# apps/market-audit-store/openapi/audit-store-v1.json
{
 "components": {
  "schemas": {
   "AccessGrant": {
    "properties": {
     "artifact_id": {
      "title": "Artifact Id",
      "type": "string"
     },
     "expires_at": {
      "format": "date-time",
      "title": "Expires At",
      "type": "string"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "artifact_id",
     "url",
     "expires_at"
    ],
    "title": "AccessGrant",
    "type": "object"
   },
   "AccessIn": {
    "additionalProperties": false,
    "properties": {
     "purpose": {
      "maxLength": 500,
      "minLength": 3,
      "title": "Purpose",
      "type": "string"
     },
     "run_id": {
      "anyOf": [
       {
        "pattern": "^run_[0-9a-f]{32}$",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Run Id"
     }
    },
    "required": [
     "purpose"
    ],
    "title": "AccessIn",
    "type": "object"
   },
   "ArtifactPrepare": {
    "additionalProperties": false,
    "properties": {
     "media_type": {
      "enum": [
       "application/json",
       "application/vnd.apache.parquet",
       "text/x-python",
       "text/csv",
       "text/plain",
       "image/png",
       "image/svg+xml",
       "application/octet-stream"
      ],
      "title": "Media Type",
      "type": "string"
     },
     "sha256": {
      "pattern": "^[0-9a-f]{64}$",
      "title": "Sha256",
      "type": "string"
     },
     "size_bytes": {
      "minimum": 0.0,
      "title": "Size Bytes",
      "type": "integer"
     },
     "upload": {
      "default": true,
      "title": "Upload",
      "type": "boolean"
     }
    },
    "required": [
     "sha256",
     "size_bytes",
     "media_type"
    ],
    "title": "ArtifactPrepare",
    "type": "object"
   },
   "ArtifactPrepared": {
    "properties": {
     "artifact_id": {
      "title": "Artifact Id",
      "type": "string"
     },
     "deduplicated": {
      "title": "Deduplicated",
      "type": "boolean"
     },
     "state": {
      "title": "State",
      "type": "string"
     },
     "upload": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/UploadTarget"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "required": [
     "artifact_id",
     "state",
     "deduplicated",
     "upload"
    ],
    "title": "ArtifactPrepared",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "description": "Contract audit-store/v1. Durable, queryable audit of AI research runs.",
  "title": "Saniti Market Audit Store",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/v1/artifacts/{artifact_id}/access": {
   "post": {
    "operationId": "access_v1_artifacts__artifact_id__access_post",
    "parameters": [
     {
      "in": "path",
      "name": "artifact_id",
      "required": true,
      "schema": {
       "pattern": "^art_[0-9a-f]{32}$",
       "title": "Artifact Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AccessIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AccessGrant"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Access",
    "tags": [
     "query"
    ]
   }
  },
  "/v1/conversations/{conversation_id}/runs": {
   "get": {
    "operationId": "get_conversation_v1_conversations__conversation_id__runs_get",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "pattern": "^[A-Za-z0-9._:-]{1,200}$",
       "title": "Conversation Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 500,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Conversation V1 Conversations  Conversation Id  Runs Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Conversation",
    "tags": [
     "query"
    ]
   }
  },
  "/v1/executions/{execution_id}/runtime": {
   "get": {
    "operationId": "get_runtime_v1_executions__execution_id__runtime_get",
    "parameters": [
     {
      "in": "path",
      "name": "execution_id",
      "required": true,
      "schema": {
       "pattern": "^exe_[0-9a-f]{24}$",
       "title": "Execution Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Runtime V1 Executions  Execution Id  Runtime Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Runtime",
    "tags": [
     "query"
    ]
   }
  },
  "/v1/internal/artifacts": {
   "post": {
    "operationId": "prepare_artifact_v1_internal_artifacts_post",
    "parameters": [
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ArtifactPrepare"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ArtifactPrepared"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Prepare Artifact",
    "tags": [
     "producer"
    ]
   }
  },
  "/v1/internal/artifacts/{artifact_id}/verify": {
   "post": {
    "operationId": "verify_artifact_v1_internal_artifacts__artifact_id__verify_post",
    "parameters": [
     {
      "in": "path",
      "name": "artifact_id",
      "required": true,
      "schema": {
       "pattern": "^art_[0-9a-f]{32}$",
       "title": "Artifact Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ArtifactVerified"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Verify Artifact",
    "tags": [
     "producer"
    ]
   }
  },
  "/v1/internal/executions": {
   "post": {
    "operationId": "register_execution_v1_internal_executions_post",
    "parameters": [
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExecutionRecord"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExecutionAck"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Register Execution",
    "tags": [
     "producer"
    ]
   }
  },
  "/v1/internal/runs": {
   "post": {
    "operationId": "register_run_v1_internal_runs_post",
    "parameters": [
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunRegistration"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunRef"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Register Run",
    "tags": [
     "producer"
    ]
   }
  },
  "/v1/internal/runs/{run_id}/artifacts": {
   "post": {
    "operationId": "link_artifacts_v1_internal_runs__run_id__artifacts_post",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "pattern": "^run_[0-9a-f]{32}$",
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LinkBatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LinkBatchAck"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Link Artifacts",
    "tags": [
     "producer"
    ]
   }
  },
  "/v1/internal/runs/{run_id}/events": {
   "post": {
    "operationId": "append_events_v1_internal_runs__run_id__events_post",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "pattern": "^run_[0-9a-f]{32}$",
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EventBatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventBatchAck"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Append Events",
    "tags": [
     "producer"
    ]
   }
  },
  "/v1/internal/runs/{run_id}/finalize": {
   "post": {
    "operationId": "finalize_v1_internal_runs__run_id__finalize_post",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "pattern": "^run_[0-9a-f]{32}$",
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Finalize"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Finalize V1 Internal Runs  Run Id  Finalize Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Finalize",
    "tags": [
     "producer"
    ]
   }
  },
  "/v1/requests/{request_id}/run": {
   "get": {
    "operationId": "get_run_by_request_v1_requests__request_id__run_get",
    "parameters": [
     {
      "in": "path",
      "name": "request_id",
      "required": true,
      "schema": {
       "pattern": "^[A-Za-z0-9._:-]{1,200}$",
       "title": "Request Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Run By Request V1 Requests  Request Id  Run Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Run By Request",
    "tags": [
     "query"
    ]
   }
  },
  "/v1/retention-holds": {
   "post": {
    "operationId": "add_hold_v1_retention_holds_post",
    "parameters": [
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/HoldIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Add Hold V1 Retention Holds Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Add Hold",
    "tags": [
     "retention"
    ]
   }
  },
  "/v1/retention-holds/{hold_id}/release": {
   "post": {
    "operationId": "release_hold_v1_retention_holds__hold_id__release_post",
    "parameters": [
     {
      "in": "path",
      "name": "hold_id",
      "required": true,
      "schema": {
       "pattern": "^hold_[0-9a-f]{32}$",
       "title": "Hold Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Release Hold V1 Retention Holds  Hold Id  Release Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Release Hold",
    "tags": [
     "retention"
    ]
   }
  },
  "/v1/retention/report": {
   "get": {
    "operationId": "retention_report_v1_retention_report_get",
    "parameters": [
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Retention Report V1 Retention Report Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Retention Report",
    "tags": [
     "retention"
    ]
   }
  },
  "/v1/runs/{run_id}": {
   "get": {
    "operationId": "get_run_v1_runs__run_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "pattern": "^run_[0-9a-f]{32}$",
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Run V1 Runs  Run Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Run",
    "tags": [
     "query"
    ]
   }
  },
  "/v1/runs/{run_id}/artifacts": {
   "get": {
    "operationId": "get_artifacts_v1_runs__run_id__artifacts_get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "pattern": "^run_[0-9a-f]{32}$",
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "in": "header",
      "name": "x-audit-accessor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Audit-Accessor"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Artifacts V1 Runs  Run Id  Artifacts Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Artifacts",
    "tags": [
     "query"
    ]
   }
  },
...
 }
}